feat: colorize only ttys
Usually kubectl is also used in aliases such as the ones
available as zsh plugins
(https://github.com/ohmyzsh/ohmyzsh/blob/master/plugins/kubectl/kubectl.plugin.zsh).

Unfortunately if you alias k to kubecolor, you won't be able
to get commands such as `kcn` to work because the output in
the subshell will be colorized as well.

This commit checks whether the user is running the command
in a TTY, and returns the colorized version only in this case,
so that kubecolor and kubectl subshell commands can coexist.