more readable colors
because strings contain the more information, and white is the most readable color,
strings are changed to white color.

table output is changed to two alternating colors, to enhance the readability of the
columns without using too many colors

key indent color alternates between yellow and cyan to enhance the indent readability
and differentiate the key from the content

while it would be great to enhance the readability of the help page, coloring
everything in yellow doesn't seem to help for that purpose. As for strings,
white is more readable and seems more suited for this complex content