color the status in yellow or red, depending on the severity
no color is used for normal status, like Running. Green is not used in order to avoid the confusion
with red color for color blind users.

yellow is used for warnings and transional status, like Pending - very useful to differentiate
Pending and Running.

red is used for failed, error, ... status